requirements.txt
mlflow>=2.21.3
pandas>=2.2.3
numpy==1.24.3
scikit-learn>=1.6.1
tensorflow-macos==2.14.0
hyperopt
